Move WaveformDisplay into separate class and add some commenting
