+++ osu.Game/osu.Game.csproj
@@ -620,6 +620,7 @@
     <Compile Include="Screens\Edit\Menus\ScreenSelectionTabControl.cs" />
     <Compile Include="Screens\Edit\Screens\EditorScreen.cs" />
     <Compile Include="Screens\Edit\Screens\Compose\Compose.cs" />
+    <Compile Include="Screens\Edit\Screens\Compose\WaveformDisplay.cs" />
     <Compile Include="Screens\Loader.cs" />
     <Compile Include="Screens\Menu\Button.cs" />
     <Compile Include="Screens\Menu\ButtonSystem.cs" />
